+++ packages/components-js/src/components/form/radio-button-wrapper/radio-button-wrapper.tsx
@@ -135,6 +135,7 @@ export class RadioButtonWrapper {
      * of the box
      */
     if (this.host.shadowRoot?.host && (event.target as HTMLElement).closest('a') === null) {
+      this.input.focus();
       this.input.click();
     }
   };